requirements.txt
streamlit==1.25.0
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.0.284
openai==0.28.1
faiss-cpu==1.7.4
tiktoken==0.4.0
